Steganography
This application lets you embed or hide important or private messages or
files into picture  files like JPG, GIF and BMP Bitmap images without
affecting the quality of actual images or files. It achieves this by
using the least significant bits of these files for embedding data which
are not used by the Image viewers or Image editors.